requirements.txt
Django==4.2.7
djangorestframework==3.14.0
mysqlclient==2.2.0
python-decouple==3.8
cryptography==41.0.7
django-cors-headers==4.3.1
Pillow==10.1.0
drf-yasg==1.21.7
